requirements.txt
streamlit
finance
requests
pandas
vaderSentiment
